backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
PyJWT>=2.9.0
bcrypt>=4.2.0
python-multipart>=0.0.12
anthropic>=0.34.2
python-dotenv>=1.0.1
pydantic[email]>=2.9.2
email-validator>=2.2.0
alembic>=1.13.3
